processor/requirements.txt
elasticsearch==6.2.0
kafka-python==1.4.2
